extract bulkSearch controller
